refactor: put buttons click functions in  different class
