+++ simple-calculator/Forms/CalForm.resx
