+++ simple-calculator/Forms/HistoryForm.resx
